# Design/middleBackendSwagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MiniMLBackend",
    "version": "1.0"
  },
  "paths": {
    "/api/CSV/csvfile/{simId}": {
      "post": {
        "tags": [
          "CSV"
        ],
        "parameters": [
          {
            "name": "simId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "CSV"
        ],
        "parameters": [
          {
            "name": "simId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}